fix bug, merge complete

+++ src/main/java/com/example/intermediate/service/CommentService.java
@@ -30,24 +30,24 @@ public class CommentService {
 
   @Transactional
   public ResponseDto<?> createComment(CommentRequestDto requestDto, Long postId, HttpServletRequest request) {
-    // header에 token이 있는지 확인
-//    if (request.getHeader("RefreshToken") == null || request.getHeader("Authorization") == null) {
-//      return ResponseDto.fail("MEMBER_NOT_FOUND",
-//          "로그인이 필요합니다.");
-//    }
+//     header에 token이 있는지 확인
+    if (request.getHeader("RefreshToken") == null || request.getHeader("Authorization") == null) {
+      return ResponseDto.fail("MEMBER_NOT_FOUND",
+          "로그인이 필요합니다.");
+    }
 
-    // refresh token이 유효한지 확인
-//    Member member = validateMember(request);
-//    if (member == null) {
-//      return ResponseDto.fail("INVALID_TOKEN", "Token이 유효하지 않습니다.");
-//    }
+//     refresh token이 유효한지 확인
+    Member member = validateMember(request);
+    if (member == null) {
+      return ResponseDto.fail("INVALID_TOKEN", "Token이 유효하지 않습니다.");
+    }
 
     Post post = postService.isPresentPost(postId);
     if (post == null) {
       return ResponseDto.fail("NOT_FOUND", "존재하지 않는 게시글입니다.");
     }
 
-    Member member = memberRepository.findById(1L).orElseGet(Member::new);
+//    Member member = memberRepository.findById(1L).orElseGet(Member::new);
 
     Comment comment = Comment.builder()
         .member(member)
@@ -67,36 +67,36 @@ public class CommentService {
 
   @Transactional
   public ResponseDto<?> updateComment(Long postId, Long commentId,  CommentRequestDto requestDto, HttpServletRequest request) {
-//    if (null == request.getHeader("Refresh-Token")) {
-//      return ResponseDto.fail("MEMBER_NOT_FOUND",
-//          "로그인이 필요합니다.");
-//    }
-//
-//    if (null == request.getHeader("Authorization")) {
-//      return ResponseDto.fail("MEMBER_NOT_FOUND",
-//          "로그인이 필요합니다.");
-//    }
-//
-//    Member member = validateMember(request);
-//    if (null == member) {
-//      return ResponseDto.fail("INVALID_TOKEN", "Token이 유효하지 않습니다.");
-//    }
+    if (null == request.getHeader("RefreshToken")) {
+      return ResponseDto.fail("MEMBER_NOT_FOUND",
+          "로그인이 필요합니다.");
+    }
+
+    if (null == request.getHeader("Authorization")) {
+      return ResponseDto.fail("MEMBER_NOT_FOUND",
+          "로그인이 필요합니다.");
+    }
+
+    Member member = validateMember(request);
+    if (null == member) {
+      return ResponseDto.fail("INVALID_TOKEN", "Token이 유효하지 않습니다.");
+    }
 
     Post post = postService.isPresentPost(postId);
     if (post == null) {
       return ResponseDto.fail("NOT_FOUND", "존재하지 않는 게시글 id 입니다.");
     }
 
-    Comment comment = isPresentComment(postId);
+    Comment comment = isPresentComment(commentId);
     if (comment == null) {
       return ResponseDto.fail("NOT_FOUND", "존재하지 않는 댓글 id 입니다.");
     }
 
-    Member member = memberRepository.findById(1L).orElseGet(Member::new);
+//    Member member = memberRepository.findById(1L).orElseGet(Member::new);
 
-//    if (comment.validateMember(member)) {
-//      return ResponseDto.fail("BAD_REQUEST", "작성자만 수정할 수 있습니다.");
-//    }
+    if (comment.validateMember(member)) {
+      return ResponseDto.fail("BAD_REQUEST", "작성자만 수정할 수 있습니다.");
+    }
 
     comment.update(requestDto);
     return ResponseDto.success(
@@ -112,29 +112,29 @@ public class CommentService {
 
   @Transactional
   public ResponseDto<?> deleteComment(Long postId, Long commentId, HttpServletRequest request) {
-//    if (null == request.getHeader("Refresh-Token")) {
-//      return ResponseDto.fail("MEMBER_NOT_FOUND",
-//          "로그인이 필요합니다.");
-//    }
-//
-//    if (null == request.getHeader("Authorization")) {
-//      return ResponseDto.fail("MEMBER_NOT_FOUND",
-//          "로그인이 필요합니다.");
-//    }
-//
-//    Member member = validateMember(request);
-//    if (null == member) {
-//      return ResponseDto.fail("INVALID_TOKEN", "Token이 유효하지 않습니다.");
-//    }
+    if (null == request.getHeader("RefreshToken")) {
+      return ResponseDto.fail("MEMBER_NOT_FOUND",
+          "로그인이 필요합니다.");
+    }
+
+    if (null == request.getHeader("Authorization")) {
+      return ResponseDto.fail("MEMBER_NOT_FOUND",
+          "로그인이 필요합니다.");
+    }
+
+    Member member = validateMember(request);
+    if (null == member) {
+      return ResponseDto.fail("INVALID_TOKEN", "Token이 유효하지 않습니다.");
+    }
 
     Comment comment = isPresentComment(commentId);
     if (null == comment) {
       return ResponseDto.fail("NOT_FOUND", "존재하지 않는 댓글 id 입니다.");
     }
 
-//    if (comment.validateMember(member)) {
-//      return ResponseDto.fail("BAD_REQUEST", "작성자만 수정할 수 있습니다.");
-//    }
+    if (comment.validateMember(member)) {
+      return ResponseDto.fail("BAD_REQUEST", "작성자만 수정할 수 있습니다.");
+    }
 
     commentRepository.delete(comment);
     return ResponseDto.success(commentId + " 번 댓글이 삭제되었습니다.");
